simplify, cleanup
- Drop initialState from the gate/model-context path; stamp the live baseline on first send instead (kept initialState as
  seed-only)
- Forward the edited message's own interactables snapshot in the edit composer so the new branch keeps its baseline
- Strip unstable_composerMetadata from the ChatModelAdapter.run context (composer-only, never model-facing)
- Add a dev-mode warning when interactable state is non-JSON (silent per-send re-stamping)
- Delete the speculative update_* tool-identity cache (no consumer depends on referential identity)
